Web-Apps/Web-App_dog_cat/requirements.txt
numpy==1.20.1
Keras==2.4.3
starlette==0.14.2
uvicorn==0.13.4
tensorflow_gpu==2.4.1
fastapi==0.68.1
Pillow==8.3.2
tensorflow==2.6.0